pyproject.toml
[project]
name = "blogstack"
version = "0.1.0"
description = ""
authors = [
    {name = "ochamekan"}
]
requires-python = ">=3.14"
dependencies = [
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "fastapi[standard] (>=0.135.3,<0.136.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-slugify (>=8.0.4,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.fastapi]
entrypoint = "src.main:app"

[tool.poetry]
package-mode = false

